created data.py
